+++ CHANGELOG.md
@@ -5,6 +5,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `Default` for `Echo`
+- `Default` for `group_concurrent::Grouped`
+- `From<S>` for `group_concurrent::Grouped<S, G: Default>`
+
 ## [0.0.99] — 2025-09-20
 
 ### Added