+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Default` for `Echo`
 - `Default` for `group_concurrent::Grouped`
 - `From<S>` for `group_concurrent::Grouped<S, G: Default>`
+- `group_concurrent::Grouped::new`
 
 ## [0.0.99] — 2025-09-20